Add new itinerary details design
